test: merge tridiag dot_general test into existing test_structure_matrix (#190 review)
The example already had `tests/usage/test_structure_matrix.py`; the new
`test_structured_matrices.py` duplicated its purpose (my PR description wrongly
claimed no test module existed). Move the batched-axis-order regression test
into the existing file and drop the duplicate.
